Use exception handling to catch failed items.
All of the methods in the pcdm classes that issue requests to the repositories now raise a pcdm.RESTAPIException if they do not get a success response from the repository. The item loading in load.py is wrapped in two levels of exception handler. The inner layer catches errors in particular items or components and attempts to rollback that particular item's transaction. The outer layer catches exceptions in creating or rolling back the transactions, and aborts the loader entirely.