LIBFCREPO-269. Add the METS metadata files to the issue.
Added as related objects of class fabio:Metadata, which then each have one pcdm:File of the actual XML. These files are represented by the class ndnp.MetadataFile, which adds fabio:MetadataDocument to their list of RDF types.

Also changed the pcdm.File constructor to take an optional title argument and use that (or the basename of the filepath) to set the dcterms:title in the file's metadata. This ensures that every pcdm:File has a title, not just ones that are created via the ndnp.File class.

https://issues.umd.edu/browse/LIBFCREPO-269