LIBFCREPO-290. Add recursive option to delete.py script.
Pass it a comma-separated list of prefixed identifiers to set the predicates for it to traverse when it does a recursive_get. The recognized prefixes are the ones defined in the pcdm module's namespace manager. Full URIs may also be given, but must be written in N3 format like <http://pcdm.org/models#hasFile>

https://issues.umd.edu/browse/LIBFCREPO-290