LIBFCREPO-510. Added LabeledThing class.
A LabeledThing is intended to be used for a hash URI resource that is either a local proxy for an authority record (in which case it has an owl:sameAs property pointing to the authority URI) or a local authority record without a URI. In both cases, an rdfs:label is required.

There is a special Place subclass of LabeledThing which also can read and return geo:lat and geo:long coordinates.

https://issues.umd.edu/browse/LIBFCREPO-510