LIBFCREPO-794. Implemented InboxWatcher to monitor for new messages
Added the "watchdog" library to "requirements.txt" and "setup.py".

Used the "watchdog" library to create InboxWatcher/InboxEventHandler
classes.

The InboxWatcher is created and started when via the "on_connected"
method of the CommandListener class. The InboxEventHandler is set up to
respond to the creation of new files, or the modification of existing
files, in the "inbox" directory by converting the file into a message
and calling the "process_message" method on CommandListener class.

The InboxWatcher is stopped when the via the "on_disconnected" method in
CommandListener.

The handling of inbox messages is somewhat changed – all messages now
call "process_message" in CommandListener via the
InboxWatcher/InboxEventHandler classes. There are two sources of
messages:

1) A file (or files) placed directly into the "inbox" directory.
2) File (or files) being send from Archelon.

In the prior implementation, the "on_message" method in CommandListener
would write out the message to a file in the "inbox" directory (via
the "add" method in the MessageBox class), and immediately process the
message. This was problematic with InboxWatcher, because the writing of
the file by CommandListener would trigger the "new file" event handler,
and the message would get processed twice. In the current
implementation, the "on_message" method only adds the file to the
"inbox" directory (via "add" method in MessageBox), and relies on the
InboxWatcher to trigger the message processing.

The startup processing of the "inbox" directory remains the same (i.e.,
in the "on_connected" method), and occurs before the InboxWatcher is
started.

In testing, when ActiveMQ is stopped, the actions taken by the
reconnection action taken in the "on_disconnected" method of
ReconnectListener appears to cause the "on_disconnected" method in
CommandListener not be to called (in which the InboxWatcher is stopped).
This was corrected by reversing the order in which the CommandListener
and ReconnectListener are added to the broker connection in
"plastron/daemon.py", so that ReconnectListener is added last.

https://issues.umd.edu/browse/LIBFCREPO-794