LIBFCREPO-821. Removed "on_modified" event handler in InboxEventHandler
When a new file is created on Linux both the "on_created" and
"on_modified" event handlers were being triggered, unlike on Mac OS X
where only the "on_created" event handler was triggered.

Since the expectation is that files in the inbox will only be created
(as opposed to modified manually in-place), the "on_modified" event
handler is redundant, and so was removed.

https://issues.umd.edu/browse/LIBFCREPO-821