LIBFCREPO-841. Handle cases when ItemLog may be None
Handle cases where ItemLog (represented by the "self.resources.completed"
variable can be None, to prevent runtime type errors.

Did this instead of setting "self.resources.completed" to an empty array
or set, because that did not seem to be in line with the use of
ItemLog as a single object.

https://issues.umd.edu/browse/LIBFCREPO-841