LIBFCREPO-839. Modified "update" command to run in daemon
Modified the "update" command to be runnable via the daemon.

Due to the presence of line breaks in the SPARQL query, it did not seem
feasible the SPARQL query via the "PlastronArg-update-query" header.

Instead, passed the SPARQL query in the message body, along with the
URIs. A typical message body looks like:

```
{
  "uri":["https://fcrepolocal/fcrepo/rest/pcdm/93/c8/cf/f5/93c8cff5-3ea3-442b-bde3-d1e493b8d998"],
  "sparql_update":"DELETE {...} INSERT {... } WHERE {}"
}
```

https://issues.umd.edu/browse/LIBFCREPO-839