LIBFCREPO-848. Fixed Plastron daemon "update" command header parsing
Corrected issue where the following "update" command arguments:

* dry-run
* validate
* use_transactions

would also be parsed as "true", even if a string such as "False" was
provided.

Modified the code to use the "strtobool" function from distutils.utils,
instead of the Python "bool" function.

Added tests for verifying header parsing of the above arguments.

https://issues.umd.edu/browse/LIBFCREPO-848