LIBFCREPO-910. Added "pycodestyle" as a "dev" dependency
The Jenkins CI server is running "pycodestyle" as part of its continuous
integration, so adding "pycodestyle" as a "dev" dependency, so that
developers can easily run pycodestyle to check for violations.

Updated the developer instructions README.md file to load the "dev"
dependencies when configuring the Python environment.

Fixed a pycodestyle violation identified by Jenkins.

https://issues.umd.edu/browse/LIBFCREPO-910