LIBFCREPO-944. Pinned "watchdog" package to v0.10.3
v0.10.4 of the "watchdog" package had an extensive update to the
"fsevent" observer, which appears to have broken functionality the
Plastron InboxWatcher relies on
(https://github.com/gorakhargosh/watchdog/pull/680).

In https://github.com/gorakhargosh/watchdog/pull/729 it is noted
hat the watchdog MacOS tests are failing, and that there are
known issues with fsevent.

The problem still exists in the latest current watchdog version
(v1.0.2), so pinning the "watchdog" package to v0.10.3 is the simplest
fix for now.

The "tests/stomp/test_inbox_watcher.py" passed when using v0.10.3, and
fails in v0.10.4 and later, so it can be used to gauge whether
"watchdog" is performing as expected by Plastron.

https://issues.umd.edu/browse/LIBFCREPO-944