LIBFCREPO-968. Modified BaseCommand to allow repository config overrides
Modified "BaseCommand" class, adding an "override_repo_config" method
to allow subclasses to override the generic repository config with a
configuration supplied by the command.

In "plastron/commands/import.py", used the "override_repo_config"
method to set the layout structure for the repository.

In "plastron/stomp/listeners.py" moved the initialization of the
"repo" Repository after calling the "override_repo_config" method for
the command, so that it will use the overridden configuration.

https://issues.umd.edu/browse/LIBFCREPO-968