LIBFCREPO-968. Fix "closed file" issue in ZipFileSource
Fixes an issue where when importing binaries from a "zip" file in the
local development environment, the import validation would succeed, but
when actually loading the files, the import process would fail.

This was occurring because in the "create" method of the "File" class
in "pcdm.py" the "digest" function in "BinaryFileSource" which read from
the source, followed by a "close" in the BinaryFileSource.__exit__
method, followed by a stream which also tried to read the source
which was now closed.

Clearing the "self.zip_file" variable in the "close" method of
"ZipFileSource" fixes the problem, as the "zip_file" variable will
get re-created when the "get_zip_file" method is via the "open" called
by the second stream.

https://issues.umd.edu/browse/LIBFCREPO-968