LIBFCREPO-972. Modified Plastron to support forwarded host
In Kubernetes, the "REST_ENDPOINT" parameter uses an internal
Kubernetes URL to communicate with umd-fcrepo-webapp. This was
causing the internal URL to be added to the fcrepo resource,
which then could not be located.

Modified the code to use a "REPO_EXTERNAL_URL" to enable "forwarding"
if the "REST_ENDPOINT" and "REPO_EXTERNAL_URL" differ in hostname
or protocol. In Kubernetes, the "REPO_EXTERNAL_URL" parameter
should be set to "FCREPO_BASE_URL".

In addition to setting the "X-Forwarded-Host" and
"X-Forwarded-Proto" headers, added functions to swap
the REST_ENDPOINT and forwarded hosts as needed
so that:

1) The forwarded host is used in the fcrepo resources
2) The REST_ENDPOINT is used to communicate with
umd-fcrepo-webapp

https://issues.umd.edu/browse/LIBFCREPO-972