LIBFCREPO-973. Use the model datatype as the default.
If a column in an import CSV has no datatype, and its corresponding property in the model class does, use the datatype defined in the model class for values in that column.

Also refactored the build_fields() function to take a model class as its second argument, instead of the property_attrs lookup dictionary. It was the only place where that dictionary was used, and it was moved inside of build_fields as an implementation detail.

https://issues.umd.edu/browse/LIBFCREPO-973