LIBFCREPO-976. Refactored auth handling into separate classes
Refactored the various forms of authorization handling into an "Auth"
class, with subclasses handling the individual implementations.

An "AuthFactory" creates and returns the proper Auth subclass for the
configuration, based on the properties in the configuration and the
authorization precedence order.

Replace code in "Repository" class and "http.py" with calls to the
"AuthFactory" and "Auth" classes, storing the Auth instance in the
repository.

https://issues.umd.edu/browse/LIBFCREPO-976