LIBFCREPO-976. Modified Auth to support "is_expired" and "refresh_auth"
Modified Auth and its subclasses to support "is_expired" and
"refresh_auth" methods.

Currently, only the "JwtSecretAuth" actually handles expiration and
refreshing.

Added "freezegun" package to "setup.py" to support testing of
expiration.

Update Repository to attempt to refresh the auth in the "request"
method, as that is the centralized point for making requests to the
server.

https://issues.umd.edu/browse/LIBFCREPO-976