LIBFCREPO-977. Modified "forwarded_endpoint" to include port
When running Plastron in Docker for the local development environment,
the "REPO_EXTERNAL_URL" parameter contains a port number (8080), that
was not being reflected in the "forwarded_endpoint" variable.

Modified the way the "forwarded_endpoint" is initialized to include
the port by using the "netloc" parameter from "urlsplit" dictionary
instead of "hostname".

https://issues.umd.edu/browse/LIBFCREPO-977