LIBFCREPO-1046. Renamed "docker-plastron.yml" file to template
Renamed "docker-plastron.yml" file to "docker-plastron-template.yml".

Did this because the file is really a template file (requires the
JWT_SECRET to be entered), and so that actual configured file
("docker-plastron.yml") can be added to the .gitignore file, so that
a fully-configured Plastron instance can be created without any
modified files showing up in "git status".

Updated the ".gitignore" file to include "docker-plastron.yml"

https://issues.umd.edu/browse/LIBFCREPO-1046