LIBFCREPO-1045. Added "contains" method to plastron.http.Repository
Added a "contains" method to the plastron.http.Repository class that
returns True if the given URI string is contained within the repository,
False otherwise. When running in Docker, the method accepts URIs
starting with either the "internal" Kubernetes URL, or the "external"
repository URL.

Modified the "get_resources" method in plastron.util.ResourceList to
use the "contains" method to determine if the resource URI is in the
repository.

https://issues.umd.edu/browse/LIBFCREPO-1045