LIBFCREPO-1076. Added "--rename-binary-column" argument to "stub"
Modified the "stub" command to take a "--rename-binary-column" optional
argument, that changes the binary_column name in the CSV output.

This is needed because the binary_column in the output will always
be a URL, while the binary_column in the input may be a filepath
(and thus have a column name that is not descriptive of the output).

Updated the CLI documentation to include the "stub" command.

https://issues.umd.edu/browse/LIBFCREPO-1076