LIBFCREPO-1084. Modified "stub" command to accept URL for "access" flag
Modified the "uri_or_curie" function in "plastron/util.py" to return
a URIRef if a simple URL is provided as an argument. This enables a
simple URL to be provided to the "access" flag of the "stub" command.

Modified the "plastron/commands/stub.py" file, so that the term returned
by by the "uri_or_curie" function is actually used to populate the
RDF type of the item being added, instead of simply using the bare
string provided in the arguments.

https://issues.umd.edu/browse/LIBFCREPO-1084