LIBFCREPO-1103. Increase test connection timeout to 10 seconds.
The 5 second timeout in the call to STOMPDaemon.stopped.wait() was yielding inconsistent results on the ci servers, suggesting that it is not guaranteed that the thread will stop within 5 seconds of the STOMP server going down. Increasing the timeout to 10 seconds in the test in the hope of having more consistent test results.

https://issues.umd.edu/browse/LIBFCREPO-1103