LIBFCREPO-1304. Build a lookup table of object class names.
In the __init_subclass__() method for RDFResourceBase, record the short name of the class being defined and a pointer to the class object in a dictionary. Then use this dictionary to look up object_class values that are given as strings.

Re-enabled skipped model tests, and added test for empty objects too.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1304