LIBFCREPO-1339. Modified vocabulary retrieval to support local testing
Modified the
`plastron-models/src/plastron/validation/vocabularies/__init__.py`
to enable the `VOCABULARIES` and `VOCABULARIES_DIR` variables to be
overridden by pytest.

The `conftest.py` method in the root project directory overrides the
`VOCABULARIES` and `VOCABULARIES_DIR` variables so that all vocabularies
used in the tests are retrieved from local file copies, instead of the
network. This makes the unit tests faster and more reliable, as they
no longer need to make a network call. It also makes it possible to
test vocabularies that have not yet been added to the vocabulary server.

Added local file copies (in "turtle" format) of all the vocabularies
used in the tests.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1339