LIBFCREPO-1425. Remove per-row publishing/hiding from Plastron import
Removed code from the
“plastron-repo/src/plastron/jobs/importjob/__init__.py” file where the
“publish” and “hidden” fields from individual CSV rows was used to
publish/hide items. This included removing code from the
“create_resource” method, which set the “umdaccess.Published” and
“umdaccess.Hidden” RDF type for newly created items (which was
problematic because it was setting the types without doing other
“publish” activities such as minting a new handle).

Updated the unit tests, and the “plastron-cli/docs/import.md” file
(which had become out-of-date with the current options provided by the
“import” command).

Based on the changes in this issue, the “import” command only publishes
via the “--publish” flag, which published all the items in the import.
There is no way to “hide” (or make visible) an item via the “import”
ommand – use the “publish” and “unpublish” commands instead (via the
“--hide” and “--visible” flags).

Added an ADR document, as the decision to remove the per-row use of the
“PUBLISH” and “HIDDEN” fields by the “import” command seemed to rise to
that level (from a philosophical/”what were we thinking” perspective).

https://umd-dit.atlassian.net/browse/LIBFCREPO-1425