LIBFCREPO-1438. Modified Import "make-templates" to handle mapped headers
Modified the "--make-template" functionality of the "import" command
to handle model HEADER_MAPs with dictionary values, by adding each
value in the dictionary as a CSV header field (regardless of the key).

This was done by introducing a "parse_model_header_map" to the
"plastron-cli/src/plastron/cli/commands/importcommand.py" file,
and checking to see if the header value is a simple string, or a
dictionary, and taking appropriate action. The "write_model_template"
was then updated to use the new method.

Added a unit tests using the "Item", "Letter", "Poster", and "Issue"
models to verify functionality. This results in coupling between the
tests and the HEADER_MAP in each of the models, but hopefully it will
be simple to update the tests as needed as the header fields evolve.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1438